+++ pkg/ipc/helper-core/helpers.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// coreimpl provide functions ...
+// Package coreimpl provides helpers functions to get certificate/key pair based on the configuration file
 package coreimpl
 
 import (
